Diplom_1 Dudin_Denis_7qa corrected
