+++ test/test_burger.py
@@ -1,34 +1,18 @@
 import pytest
-from unittest.mock import Mock
 
-
-class MockBun:
-    mock_bun = Mock()
-    mock_bun.get_name.return_value = "Краторная булка"
-    mock_bun.get_price.return_value = 1255
-
-
-class MockIngredientFilling:
-    mock_ingredient_filling = Mock()
-    mock_ingredient_filling.get_name.return_value = "Говяжий метеорит"
-    mock_ingredient_filling.get_price.return_value = 3000
-    mock_ingredient_filling.get_type.return_value = "FILLING"
-
-
-class MockIngredientSauce:
-    mock_ingredient_sauce = Mock()
-    mock_ingredient_sauce.get_name.return_value = "Соус Spicy-X"
-    mock_ingredient_sauce.get_price.return_value = 90
-    mock_ingredient_sauce.get_type.return_value = "SAUCE"
+from data import MockBun, MockIngredientFilling, MockIngredientSauce, ReceiptBurger
 
 
 class TestBurger:
     def test_burger_set_bun(self, burger):
+        burger.set_buns(MockBun.mock_bun)
         assert burger.get_price() == 2510
 
     @pytest.mark.parametrize("ingredient, total_price",
-                             [[MockIngredientSauce.mock_ingredient_sauce, 2600],
-                              [MockIngredientFilling.mock_ingredient_filling, 5510]])
+                             [
+                                 (MockIngredientFilling.mock_ingredient_filling, 5510),
+                                 (MockIngredientSauce.mock_ingredient_sauce, 2600)
+                             ])
     def test_burger_add_ingredient(self, burger, ingredient, total_price):
         burger.add_ingredient(ingredient)
         assert burger.get_price() == total_price
@@ -47,5 +31,4 @@ class TestBurger:
     def test_burger_get_receipt(self, burger):
         burger.add_ingredient(MockIngredientFilling.mock_ingredient_filling)
         receipt = burger.get_receipt()
-        assert "(==== Краторная булка ====)" in receipt
-        assert "= filling Говяжий метеорит =" in receipt
+        assert receipt == ReceiptBurger.receipt_burger
